Revert "Kernel+SystemServer: Make KCOVDevice a character device"
This reverts commit 9dbec601b0fdfe18ec425a6dd6518740309fd78a.

This change was nonsense. For KCOV to be performant (or at least not
even slower) we need to mmap the PC buffer from both user and kernel
space at the same time. You can't mmap a character device.

Plus even if we did invent a new method to exfiltrate the coverage
information out of the kernel, it would be incompatible with existing
kernel fuzzers. That would be kind of annoying. 🙃